Include user agent in the banned IP email
We often receive these email notifications about a banned IP. 
Sometimes we detect a hoster and block these. But sometimes it comes from a regular broadband provider and we can't really understand more about the request. Including the user agent to maybe learn more about the request and whether we can potentially ban / block this user agent in the future if it's not a regular browser.